feat(DPLAN-17512): adjust the unsavedChangesGuardMixin.js: remove globalConfirm.js file and move the logic into mixin
